PersonalSettings refactoring
